WaifuGrabber/requirements.txt
flask
nekos.py
